move date toInstant util to common utils dir
